Break logic out into utility classes/modules
